+++ openmetadata-ui/src/main/resources/ui/src/styles/position.less
@@ -19,6 +19,6 @@
   display: flex;
 }
 
-flex-1 {
+.flex-1 {
   flex: 1;
 }
